refactor: 수정 - 분리된 유틸들, 공통 파일로 묶음
