+++ client/src/component/Navigator.js
@@ -1,6 +1,5 @@
 import Component from '@component/Component.js';
-import { templateToElementNodes } from '@utils/generateElement.js';
-import { appendChildAll } from '@utils/document.js';
+import { appendChildAll, templateToElementNodes } from '@utils/document.js';
 
 // eslint-disable-next-line
 import style from '@stylesheet/component/Navigator.scss';
