+++ client/src/utils/helper.js
@@ -1,3 +1,8 @@
+export const numberFormat = (number) => {
+  const intl = Intl.NumberFormat('en-US');
+  return intl.format(number);
+};
+
 export const getCurrentDatetime = () => {
   const date = analyzeDatetime();
   return `${date.year}-${date.month}-${date.date} ${date.hour}:${date.minute}:${date.second}`;
@@ -27,3 +32,18 @@ export const analyzeDatetime = (datetimeStr) => {
     second,
   };
 };
+
+/**
+ * Generates a GUID string.
+ * @returns {string} The generated GUID.
+ * @example af8a8416-6e18-a307-bd9c-f2c947bbb3aa
+ * @author Slavik Meltser.
+ * @link http://slavik.meltser.info/?p=142
+ */
+export function guid() {
+  function _p8(s) {
+    var p = (Math.random().toString(16) + '000000000').substr(2, 8);
+    return s ? '-' + p.substr(0, 4) + '-' + p.substr(4, 4) : p;
+  }
+  return _p8() + _p8(true) + _p8(true) + _p8();
+}
